Change clean-up to only affect current job files

+++ app/src/file_management.py
@@ -1,6 +1,7 @@
 import hashlib
 import logging
 import os
+from typing import List
 from zipfile import ZipFile
 
 from requests import get
@@ -10,31 +11,26 @@ zips_folder = "_zips"
 data_folders = (pdfs_folder, zips_folder)
 
 
-def urls_to_zip(pdf_urls):
+def urls_to_zip(pdf_urls: List[str], job_identifier: str):
     if not os.path.exists(pdfs_folder):
         os.mkdir(pdfs_folder)
     if not os.path.exists(zips_folder):
         os.mkdir(zips_folder)
 
-    source_folder: str = _process_files(pdf_urls)
-    zipped_file: str = _zip_files(source_folder)
+    source_folder: str = _process_files(pdf_urls, job_identifier)
+    zipped_file: str = _zip_files(source_folder, job_identifier)
 
     return zipped_file
 
 
-def _process_files(all_pdfs) -> str:
+def _process_files(all_pdfs: List[str], job_identifier: str) -> str:
     files_count: int = len(all_pdfs)
 
     logging.info("#" * 80)
     logging.info(f"PROCESSING {files_count} FILES")
     logging.info("#" * 80)
 
-    h = hashlib.new("sha256")
-    pdfs_encoding_name: str = "-".join(sorted(all_pdfs))
-    h.update(pdfs_encoding_name.encode("utf-8"))
-    hex_digest: str = h.hexdigest()
-
-    source_folder = _archive_or_create_folder(hex_digest)
+    source_folder = _archive_or_create_folder(job_identifier)
 
     for index, pdf in enumerate(all_pdfs):
         file_name: str = "-".join(pdf.split("/")[-3:])
@@ -57,8 +53,8 @@ def _process_files(all_pdfs) -> str:
     return source_folder
 
 
-def _archive_or_create_folder(hex_digest) -> str:
-    source_folder: str = os.path.join(pdfs_folder, hex_digest)
+def _archive_or_create_folder(job_identifier: str) -> str:
+    source_folder: str = os.path.join(pdfs_folder, job_identifier)
 
     if not os.path.exists(pdfs_folder):
         os.mkdir(pdfs_folder)
@@ -69,9 +65,8 @@ def _archive_or_create_folder(hex_digest) -> str:
     return source_folder
 
 
-def _zip_files(source_folder) -> str:
-    file_hash: str = source_folder.split("/")[-1]
-    zip_file_name: str = os.path.join(zips_folder, f"pdfs-{file_hash}.zip")
+def _zip_files(source_folder: str, job_identifier: str) -> str:
+    zip_file_name: str = os.path.join(zips_folder, f"pdfs-{job_identifier}.zip")
     with ZipFile(zip_file_name, "w") as zip_obj:
         for folder_name, _, filenames in os.walk(source_folder):
             for filename in filenames:
@@ -81,17 +76,33 @@ def _zip_files(source_folder) -> str:
     return zip_file_name
 
 
-def clean_up() -> None:
+def encode_file_names(all_pdfs: List[str]) -> str:
+    h = hashlib.new("sha256")
+
+    pdfs_encoding_name: str = "-".join(sorted(all_pdfs))
+    h.update(pdfs_encoding_name.encode("utf-8"))
+    hex_digest: str = h.hexdigest()
+
+    return hex_digest
+
+
+def clean_up_after_current_job(job_identifier: str) -> None:
     for folder in data_folders:
-        _remove_folder_and_contents(folder)
+        _remove_folder_and_contents(folder, job_identifier)
 
 
-def _remove_folder_and_contents(folder_path) -> None:
+def _remove_folder_and_contents(folder_path: str, job_identifier: str) -> None:
     for folder_name, subfolders, filenames in os.walk(folder_path):
         if len(filenames) == 0:
             continue
 
+        if job_identifier not in folder_name:
+            continue
+
         for filename in filenames:
+            if job_identifier not in filename:
+                continue
+
             file_path: str = os.path.join(folder_name, filename)
             logging.info(f"Removing {file_path}")
             os.remove(file_path)
